+++ demo/components/docs/ExampleRaw.vue
@@ -42,7 +42,7 @@
       <li class="vsm-section vsm-mob-hide">
         <button>My Button</button>
       </li>
-      <!--Display when user media screen below from $vsm-media (scss)-->
+      <!--Set "display: block" for the .vsm-mob-show class to display content-->
       <vsm-mob>Mobile Content</vsm-mob>
     </template>
   </vsm-menu>