+++ demo/components/docs/ExampleRaw.vue
@@ -22,10 +22,7 @@
     </template>
     <template #before-nav>
       <!--Image or svg of website logo-->
-      <li
-        class="vsm-section logo-section"
-        style="width: 50px; height: 50px"
-      >
+      <li style="width: 50px; height: 50px">
         <img
           src="https://vuejs.org/images/logo.png"
           alt="My Logo"
@@ -38,8 +35,7 @@
     </template>
     <template #after-nav>
       <!--Mobile Burger, buttons, etc-->
-      <!--For the same styles - add the vsm-section-->
-      <li class="vsm-section vsm-mob-hide">
+      <li class="vsm-mob-hide">
         <button>My Button</button>
       </li>
       <!--Set "display: block" for the .vsm-mob-show class to display content-->
@@ -57,7 +53,7 @@
  *
  * After #after-nav and #before-nav it is recommended to use
  * to maintain the correct HTML structure:
- *   <li class="vsm-section">
+ *   <li><!--Content--></li>
  */
 
 export default {